not entering walls!!!!

+++ www/TileMap.h
@@ -54,7 +54,7 @@ public:
 
 	int getTileSize() const { return tileSize; }
 
-	glm::ivec2 getMinimumTranslationVector(const glm::ivec2 &playerPos, const glm::ivec2 &playerSize) const;
+	glm::vec2 getMinimumTranslationVector(const glm::ivec2 &playerPos, const glm::ivec2 &playerSize) const;
 	bool collisionMoveLeft(const glm::ivec2 &pos, const glm::ivec2 &size) const;
 	bool collisionMoveRight(const glm::ivec2 &pos, const glm::ivec2 &size) const;
 	bool collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, int *posY) const;
